utest updates to accomodate for functions moving around
